refactor middleware names
