+++ server/api.go
@@ -19,8 +19,8 @@ func getRoutes() []func() chi.Router {
 func getMiddleware() []func(http.Handler) http.Handler {
 	args := serverutil.GetArgs()
 	return []func(http.Handler) http.Handler{
-		middleware.LoggerMiddleware(),
-		middleware.AuthenticationMiddleware(middleware.AuthenticationMiddlewareOptions{
+		middleware.LoggerMiddlewareFactory(),
+		middleware.AuthenticationMiddlewareFactory(middleware.AuthenticationMiddlewareOptions{
 			Domain:   args.Auth0Domain,
 			Audience: args.Auth0Audience,
 		}),
